Interact
I replaced the gun model with my pickaxe model and rewrote the code from Shoot to Interact. This way, obstacles can be programmed to specific interactions based on their type later.  Writing of type, I also added the BlockData to the boulder, since it is useful for adding, well, data to the objects.