Treasure
I'm working on the treasure implementation. I've already made the map creator able to replace blocks to place the treasures, and I've managed to implement a DPS for checking if the treasure is reachable. Currently, I'm trying to write the code necessary for creating paths, when a path isn't available. With mixed success.