Fixed
And thanks to Esben, I managed to fix the problem. Apparently, the C# Random class uses time as a seed for random generation, so creating a new object each time I need a random double would result in the same number, when I ran it through as fast as I did. Now it works, so it's time to move on!