Debugging
I've added the first 6/24 filters, and the computation time for the ANN is horrible now. I've added some debugging using Stopwatch (thanks Esben) and I need more of it now. Along with that I'll also do some proper testing to and experimentation to find the optimal solution to this problem. And I should probably also debug the whole ANN to check for bugs!