MLDebugger
I've made a static class in the ML namescape for debugging. I've already reworked it, so it prints every detail from the run in one line, which is styled a bit for readability. Next step for this class will be multiple levels of debug detail, which can be changed by the user at will, but this will probably be annoying to make.
For now, this should do for some performance testing for the thesis!