Serilization Done
After remaking some of the current serialization progress, I've finally found the solution I needed. It works with deserilization as well, I can save and load as I please, and I can still run the network afterwards (as long as the number of inputs doesn't change from the previous run, which it in reality shouldn't ever do). Lastly, I've made it saveable to the CNNSaver, and I've made the CNNSaver's string into a giant textarea, so it is easier to read in the inspector. What a nightmare this has been... But it is done!