Refactors
I've made some edits to the CNN methods for Training, Running, and Fully Connected, so they give more control. I've also added a new function to pass an ANN to the CNN, so a custom one can be used. I can imagine the problems that will occur later, because of this choice...