Debugging
Even more debugging, this time by printing EVERYTHING from the network as it propagates forth and back. And then I've gone through the numbers and checked the values. It resulted in some errors for printing, but also noticeable changes to the network itself (e.g.: the pooling mask is now a float? instead float)