-30 IQ Points
I really should be more careful with my code when working with this stuff. I was using the same delta value for calculating the filter delta and for setting into my error map during the Conv Backprop. This led to an exponentially growing delta being added to the error map. Now I've added a current delta (for the error map) and a total delta (for the current filter pixel). I've spent so much time on this, so let this be an important lesson!