Training Ready
The networks are ready for training, and I've started the implementation of the FSM, though it's nowhere near ready or anything. But I would like to train the networks and make some initial comparisons. I'll probably do the training at the desktop, as it is faster. And maybe I can look into emulating the movement easily instead of playback?