BT Works
Yes. It works. It's not smart. It's not good. But it works. Now I just need some minor upgrades to its decision making and perhaps test a longer delay for the agent's moves, as it seems to overshoot from time to time, which kills it. But I can move on, though I'd prefer to optimize the BT and FSM a bit more first, but time is running out...